(Daniel) nicer way to incorporate networks, camera, and code logic together for physical experiments.
Still TODO:

(1) we have to keep deleting images for the directory, it's a bit annoying.

(2) Exit condition is going to be when the human hits ESC key when we've got sufficient coverage. It would be better to do off the shelf contour detection code to automatically query the height.

But otherwise, the whole 'hitting enter' for the camera, I think I can do that with deft tab manipulation, just need someone else to video I think